Add a brake to the switch
I didn't quite realize, why it is important to add a break at the very end of a switch.
But I have to admit, that it seems legit.
One could think: "Oh, theres a switch! I'll add my case right under the last one!" and might forget about the break.

resolves #3 A switch statement does not contain a break